// src/app/interfaces/contractor.ts
export interface Respuesta {
  id: number;
  first_name: string;
  last_name: string;
  email: string;
  type: string;
  phone_number: string;
  profile_image?: any;
  address?: any;
  email_verified_at?: any;
  rating?: any;
  confirmed: string;
  active: string;
  created_at: string;
  updated_at: string;
}